+++ signin/signin.html
@@ -66,6 +66,6 @@
         </div>
       </div>
     </div>
-    <script src="/script/signin.js"></script>
+    <script type="module" src="/script/signin.js"></script>
   </body>
 </html>
