+++ signup/signup.html
@@ -54,7 +54,11 @@
           <div class="sign-input-box sign-password">
             <label class="sign-input-label">비밀번호 확인</label>
             <div>
-              <input class="sign-input check" id="password-input-check" type="password" />
+              <input
+                class="sign-input check"
+                id="password-input-check"
+                type="password"
+              />
               <button class="eye-button" type="button">
                 <img src="/images/eye-off.svg" class="eye-img" id="eye-img2" />
               </button>
@@ -78,6 +82,6 @@
         </div>
       </div>
     </div>
- <script src="/script/signup.js"></script>
+    <script type="module" src="/script/signup.js"></script>
   </body>
 </html>
